% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{adaptive_reweighting}
\alias{adaptive_reweighting}
\title{Select cells using adaptive reweighting.}
\usage{
adaptive_reweighting(
  seu,
  markers,
  n_requested_cells,
  mode = "marker_based",
  n_of_pcs = 30,
  knn = 20,
  res = 0.8
)
}
\arguments{
\item{seu}{A processed \code{Seurat} object (must be normalized, the most variable features should have already been found using \code{FindVariableFeatures} and should be scaled using \code{ScaleData})}

\item{markers}{A \code{yaml} file with the set of markers to be used. The name of the marker must match that in \code{seu}, else an error is thrown.}

\item{n_requested_cells}{Number of cells to sample from dataset. Cannot be bigger than total number of cells in \code{seu}.}

\item{mode}{Either \code{marker_based} or \code{cluster_based}. When marker_based is selected, cells are sampled from aggregated clusters that have been putatively assigned to a cell type based on the average gene expression. When cluster_based is used cells are sampled from the individual clusters.}

\item{n_of_pcs}{Defines the number of PCA dimensions to use for clustering. Defaults to 30.}

\item{knn}{Defines the number of k for the k-nearest neighbor algorithm during clustering. Defaults to 20.}

\item{res}{Defines the clustering resulution. Defaults to 0.8.}
}
\value{
None
}
\description{
Select cells using adaptive reweighting.
}
