% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{cluster_enrich}
\alias{cluster_enrich}
\title{calculates the average expression of marker genes in each cluster}
\usage{
cluster_enrich(markers, expression)
}
\arguments{
\item{markers}{an object with cell type markers read in by get_markers()}

\item{expression}{dataframe with expression values for all cells and marker genes. rows are cells, columns are markers. The first column should be \code{cell_id}.}
}
\description{
calculates the average expression of marker genes in each cluster
}
