% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_markers}
\alias{get_markers}
\title{Read in user specified marker file.}
\usage{
get_markers(yaml_path)
}
\arguments{
\item{yaml_path}{Path to a yaml file. The file extension must be yml}
}
\value{
A list of positive and negative markers for each cell type
}
\description{
Read in user specified marker file.
}
