% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_aliases.R
\docType{data}
\name{io}
\alias{io}
\title{International organizations list}
\format{
An object of class \code{character} of length 31.
}
\usage{
io
}
\description{
A character vector of international organization names and acronyms
used for entity detection across the package.
}
\keyword{internal}
