% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jfk19610120}
\alias{jfk19610120}
\title{JFK Speech: Inaugural Address (January 20, 1961)}
\format{
A character string containing the full text of the speech.
}
\source{
\url{https://www.govinfo.gov/app/details/PPP-1961-book1}.
These original source texts are U.S. government works and are not subject to
copyright protection in the United States under 17 U.S.C. § 105.
The package authors do not claim copyright in the original presidential speech texts.
}
\usage{
jfk19610120
}
\description{
Full text of JFK's Inaugural Address on January 20, 1961.
}
\keyword{datasets}
