% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jfk19630610}
\alias{jfk19630610}
\title{JFK Speech: Commencement Address at American University (June 10, 1963)}
\format{
A character string containing the full text of the speech.
}
\source{
\url{https://www.govinfo.gov/app/details/PPP-1963-book1}.
These original source texts are U.S. government works and are not subject to
copyright protection in the United States under 17 U.S.C. § 105.
The package authors do not claim copyright in the original presidential speech texts.
}
\usage{
jfk19630610
}
\description{
Full text of Senator John F. Kennedy's commencement address at American University on June 10, 1963.
}
\keyword{datasets}
