% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexity.R
\name{prep_lc_regex}
\alias{prep_lc_regex}
\title{Prepare low-complexity regex pattern}
\usage{
prep_lc_regex(items)
}
\arguments{
\item{items}{Character vector of low-complexity words and phrases.}
}
\value{
A single regex string.
}
\description{
Prepare low-complexity regex pattern
}
\keyword{internal}
