% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexity.R
\name{prep_phrase_regex}
\alias{prep_phrase_regex}
\title{Prepare phrase regex pattern}
\usage{
prep_phrase_regex(phrases)
}
\arguments{
\item{phrases}{Character vector of phrases.}
}
\value{
A single regex string.
}
\description{
Prepare phrase regex pattern
}
\keyword{internal}
