% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexity.R
\name{prep_wordstem_regex}
\alias{prep_wordstem_regex}
\title{Prepare word-stem regex pattern}
\usage{
prep_wordstem_regex(stems)
}
\arguments{
\item{stems}{Character vector of word stems.}
}
\value{
A single regex string.
}
\description{
Prepare word-stem regex pattern
}
\keyword{internal}
