% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{reweighting}
\alias{reweighting}
\title{Runs geneset enrichment analysis to identify overlapping clusters}
\usage{
reweighting(seu, markers, cluster_output)
}
\arguments{
\item{seu}{a seurat object}

\item{markers}{an object with cell type markers read in by get_markers()}

\item{cluster_output}{}
}
\description{
Runs geneset enrichment analysis to identify overlapping clusters
}
