% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{sample_cells}
\alias{sample_cells}
\title{Samples the requested number of cells from identified clusters}
\usage{
sample_cells(cluster_output, n_requested_cells)
}
\arguments{
\item{cluster_output}{cell id - cluster/putative cell type mapping dataframe}

\item{n_requested_cells}{Number of cells to select}
}
\description{
Samples the requested number of cells from identified clusters
}
