% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{seurat_cluster}
\alias{seurat_cluster}
\title{Creates PCA embedding if missing, clusters cells and returns a cluster - cell id mapping.}
\usage{
seurat_cluster(seu, n_of_pcs, knn, res)
}
\value{
A dataframe with two columns: cell_id and cluster_id for each cell.
}
\description{
Creates PCA embedding if missing, clusters cells and returns a cluster - cell id mapping.
}
