% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leadeR-package.R
\docType{package}
\name{leadeR-package}
\alias{leadeR}
\alias{leadeR-package}
\title{leadeR: Profiling Leaders at a Distance}
\description{
The leadeR package profiles political leaders using text analysis,
implementing Leadership Trait Analysis (LTA) and Operational Code
Analysis (OCA). You provide text data and the package performs the
analyses.
}
\author{
\strong{Maintainer}: Mitsuru Mukaigawara \email{mitsuru_mukaigawara@g.harvard.edu} (\href{https://orcid.org/0000-0001-6530-2083}{ORCID})

Authors:
\itemize{
  \item Joshua Kertzer \email{jkertzer@gov.harvard.edu} (\href{https://orcid.org/0000-0002-7358-0638}{ORCID})
}

}
\keyword{internal}
